+++ source/patterns/@aws-solutions-constructs/aws-iot-sqs/test/integ.dead-letter-queue-off.ts
@@ -26,7 +26,7 @@ stack.templateOptions.description = 'Integration Test for aws-iot-sqs';
 const props: IotToSqsProps = {
   deployDeadLetterQueue: false,
   queueProps: {
-    queueName: 'queue-name'
+    queueName: 'dlq-off-test-queue-name'
   },
   iotTopicRuleProps: {
     topicRulePayload: {