+++ source/patterns/@aws-solutions-constructs/core/test/s3-bucket-helper.test.ts
@@ -12,7 +12,7 @@
  */
 
 import { expect as expectCDK, haveResource, ResourcePart } from '@aws-cdk/assert';
-import { Duration, Stack } from '@aws-cdk/core';
+import { Duration, Stack, RemovalPolicy } from '@aws-cdk/core';
 import * as s3 from '@aws-cdk/aws-s3';
 import * as s3n from '@aws-cdk/aws-s3-notifications';
 import * as sqs from '@aws-cdk/aws-sqs';
@@ -55,6 +55,33 @@ test('s3 bucket with bucketProps', () => {
 
   expectCDK(stack).to(haveResource("AWS::S3::Bucket", {
     BucketName: "mybucket"
+  }));
+});
+
+test('s3 bucket with default props', () => {
+  const stack = new Stack();
+
+  defaults.buildS3Bucket(stack, {});
+
+  expectCDK(stack).to(haveResource("AWS::S3::Bucket", {
+    BucketEncryption: {
+      ServerSideEncryptionConfiguration: [
+        {
+          ServerSideEncryptionByDefault: {
+            SSEAlgorithm: "AES256"
+          }
+        }
+      ]
+    },
+    PublicAccessBlockConfiguration: {
+      BlockPublicAcls: true,
+      BlockPublicPolicy: true,
+      IgnorePublicAcls: true,
+      RestrictPublicBuckets: true
+    },
+    VersioningConfiguration: {
+      Status: "Enabled"
+    }
   }));
 });
 
@@ -236,6 +263,33 @@ test('s3 bucket versioning turned off', () => {
       BlockPublicPolicy: true,
       IgnorePublicAcls: true,
       RestrictPublicBuckets: true
+    }
+  }));
+});
+
+test('s3 bucket with LoggingBucket and auto delete objects', () => {
+  const stack = new Stack();
+
+  defaults.buildS3Bucket(stack, {
+    loggingBucketProps: {
+      removalPolicy: RemovalPolicy.DESTROY,
+      autoDeleteObjects: true
+    }
+  });
+
+  expectCDK(stack).to(haveResource("AWS::S3::Bucket", {
+    AccessControl: "LogDeliveryWrite"
+  }));
+
+  expectCDK(stack).to(haveResource("Custom::S3AutoDeleteObjects", {
+    ServiceToken: {
+      "Fn::GetAtt": [
+        "CustomS3AutoDeleteObjectsCustomResourceProviderHandler9D90184F",
+        "Arn"
+      ]
+    },
+    BucketName: {
+      Ref: "S3LoggingBucket800A2B27"
     }
   }));
 });